fix: unmounting error is resolved with return function in hook cleanup: moved getHatchProgress to useHatchProgress
